Turn on multiple bits. Change uint8_t to int.

+++ src/led_driver.h
@@ -2,7 +2,7 @@
 #define _LED_DRIVE_H
 
 void led_init(uint16_t* leds);
-void led_turn_on(uint8_t lednum);
-void led_turn_off(uint8_t lednum);
+void led_turn_on(int lednum);
+void led_turn_off(int lednum);
 
 #endif /* _LED_DRIVE_H */
